frefactor(services): logic battle, logic pokemon team, logic battle
