+++ backend/battles/forms.py
@@ -3,7 +3,7 @@ from django.forms import ModelChoiceField, ModelForm, ValidationError
 from pokemons.models import Pokemon
 
 from .models import Battle, Team, TeamPokemon
-from .pokemon import check_valid_team
+from .services.logic_team_pokemon import check_valid_team
 
 
 class BattleForm(ModelForm):
