+++ backend/battles/services/api_integration.py
@@ -39,16 +39,3 @@ def get_pokemon(poke_id):
         get_pokemon_api(poke_id)
         return Pokemon.objects.get(poke_id=poke_id)
     return pokemon
-
-
-def sum_points(pokemons):
-    points = 0
-    for pokemon in pokemons:
-        points += pokemon.attack + pokemon.defense + pokemon.hp
-    print(points)
-    return points
-
-
-def check_valid_team(pokemons):
-    is_valid = sum_points(pokemons) <= 600
-    return is_valid
